Different Classes for Login and Register form
